WIP: Convert .m2 file to annotated text.
This generates a sentence-split annotated corpus which may be more
convenient to work with in some applications. There are some caveats:

1. Sentences are re-aligned with Errant, so annotation spans are
   different from the original annotated text.
2. Error types are copied from the original texts using a set of
   heuristics that are not perfect.
3. The sentences are word-tokenized, meaning the text looks like this:
   "Hello , world !". It's an artifact of aligning with Errant.

In future, we may implement a better way to split the corpus into
sentences while keeping original annotations. For now, this is enough to
enable applications like automatic error type classification.

TODO: Several sentences are skipped for weird reasons.

TODO: Split into individual documents and expose in the Python package.